moved nats client module out from lib
