+++ nest/global/presence/client/presence-client.module.ts
@@ -1,6 +1,6 @@
 import { Module }         from '@nestjs/common'
 import { PresenceClient } from './presence.client'
-import { ClientModule }   from '../../../lib/client/client.module'
+import { ClientModule }   from '../../../client/client.module'
 
 @Module({
     imports  : [ClientModule],
