+++ nest/global/presence/emitter/presence-emitter.module.ts
@@ -1,6 +1,6 @@
 import { Module }          from '@nestjs/common'
 import { PresenceEmitter } from './presence.emitter'
-import { ClientModule }    from '../../../lib/client/client.module'
+import { ClientModule }    from '../../../client/client.module'
 
 @Module({
     imports  : [
