+++ nest/local/character/client/character-client.module.ts
@@ -1,6 +1,6 @@
 import { Module }          from '@nestjs/common'
 import { CharacterClient } from './character.client'
-import { ClientModule }    from '../../../lib/client/client.module'
+import { ClientModule }    from '../../../client/client.module'
 
 @Module({
     imports  : [ClientModule],
